+++ main.rb
@@ -11,6 +11,7 @@ require "mechanize"
 require "nokogiri"
 require 'net/http'
 require 'open-uri'
+require_relative 'db_query'
 #Need nokogir?
 agent = Mechanize.new
 
@@ -23,15 +24,18 @@ country_code = end_link.gsub(/[^\d]/,"")
 country = end_link.split("-")[1] #change country name to proper TA's country format to prevent redirects
 
 city_links = get_city_links(country, country_code)
+
 city_names = get_city_names(city_links)
 geocodes = get_city_geocodes(city_links)
 
-restaurant_links = get_restaurant_urls(city_links, geocodes, city_names)
+restaurant_links = get_restaurant_urls(geocodes, city_names)
 
 
 restaurant_links.each do |key, array|
   array.length.times do |index|
+
     restaurant_info= scrape_restaurant_page(key,"https://www.tripadvisor.ca/#{array[index]}")
+    puts restaurant_info['name']
     Restaurant.create(
     name: restaurant_info['name'],
     city: restaurant_info['city'],
@@ -40,3 +44,4 @@ restaurant_links.each do |key, array|
     )
   end
 end
+db_query
