+++ rank.rb
@@ -1,0 +1,9 @@
+def ci_lower_bound(pos, n, confidence)
+  if n == 0
+    return 0
+  end
+  z = Statistics2.pnormaldist(1-(1-confidence)/2)
+  phat = 1.0*pos/n
+  value = (phat + z*z/(2*n) - z * Math.sqrt((phat*(1-phat)+z*z/(4*n))/n))/(1+z*z/n)
+  value*10**8
+end
